+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Stochastic tester [MR22](https://github.com/Qrlew/pyqrlew/pull/22)
+
 ## [0.4.5] - 2023-10-29
 ### Changed
 - Use the last version of qrlew